requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiofiles=23.2.1=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
anndata=0.10.5.post1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
array-api-compat=1.4.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
asyncer=0.0.2=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.3=pyha770c72_0
bidict=0.22.1=pypi_0
biopython=1.83=py310hd125d64_0
bs4=4.12.3=hd8ed1ab_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2023.11.17=hf0a4a13_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
chainlit=1.0.200=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
crewai=0.28.8=pypi_0
crewai-tools=0.1.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cssselect=1.2.0=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.5.14=pypi_0
debugpy=1.8.0=py310h1253130_1
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
distlib=0.3.8=pypi_0
distro=1.9.0=pypi_0
dnspython=2.5.0=pypi_0
docopt=0.6.2=pypi_0
eutils=0.6.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
faiss-cpu=1.7.4=pypi_0
fastapi=0.108.0=pypi_0
fastapi-socketio=0.0.10=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=pypi_0
filetype=1.2.0=pypi_0
fonttools=4.47.2=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
habanero=1.2.3=pypi_0
html2text=2020.1.16=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.20.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.6=pypi_0
igraph=0.11.3=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib_metadata=7.0.1=hd8ed1ab_0
iniconfig=2.0.0=pypi_0
ipykernel=6.29.0=pyh3cd1d5f_0
ipython=8.20.0=pyh707e725_0
ipywidgets=8.1.1=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpickle=3.0.2=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py310hbe9552e_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.1.5=pypi_0
langchain-community=0.0.17=pypi_0
langchain-core=0.1.18=pypi_0
langchain-openai=0.0.5=pypi_0
langchainhub=0.1.14=pypi_0
langsmith=0.0.83=pypi_0
lark=1.1.9=pypi_0
lazify=0.4.0=pypi_0
leidenalg=0.10.2=pypi_0
levenshtein=0.23.0=pypi_0
libblas=3.9.0=21_osxarm64_openblas
libcblas=3.9.0=21_osxarm64_openblas
libcxx=16.0.6=h4653b0c_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_2
libgfortran5=13.2.0=hf226fd6_2
liblapack=3.9.0=21_osxarm64_openblas
libopenblas=0.3.26=openmp_h6c19121_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.44.2=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
literalai=0.0.103=pypi_0
llvm-openmp=17.0.6=hcd81f8e_0
llvmlite=0.42.0=pypi_0
loguru=0.7.2=pypi_0
lxml=5.1.0=pypi_0
markupsafe=2.1.4=pypi_0
marshmallow=3.20.2=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
metapub=0.5.5=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
natsort=8.4.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h463b476_2
neo4j=5.17.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numba=0.59.0=pypi_0
numexpr=2.9.0=pypi_0
numpy=1.26.3=py310hd45542a_0
openai=1.11.0=pypi_0
openssl=3.2.0=h0d3ecfb_1
opentelemetry-api=1.22.0=pypi_0
opentelemetry-exporter-otlp=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.22.0=pypi_0
opentelemetry-instrumentation=0.43b0=pypi_0
opentelemetry-proto=1.22.0=pypi_0
opentelemetry-sdk=1.22.0=pypi_0
opentelemetry-semantic-conventions=0.43b0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.0=py310h401b61c_0
paper-qa=4.0.0rc7=pypi_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pinecone-client=2.2.2=pypi_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pyhd8ed1ab_0
plotly=5.18.0=pypi_0
pluggy=1.4.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.2=pypi_0
psutil=5.9.8=py310hd125d64_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybiopax=0.1.4=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pynndescent=0.5.11=pypi_0
pyparsing=3.1.1=pypi_0
pypdf=4.0.0=pypi_0
pyproject-api=1.6.1=pypi_0
pytest=7.4.4=pypi_0
python=3.10.13=h2469fbe_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-engineio=4.8.2=pypi_0
python-graphql-client=0.4.3=pypi_0
python-levenshtein=0.23.0=pypi_0
python-multipart=0.0.6=pypi_0
python-socketio=5.11.0=pypi_0
python-tzdata=2023.4=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py310hbb13138_0
rapidfuzz=3.6.1=pypi_0
readline=8.2=h92ec313_1
referencing=0.33.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rpds-py=0.17.1=pypi_0
safetensors=0.4.2=pypi_0
scanpy=1.9.8=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
session-info=1.0.0=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
simple-websocket=1.0.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pyhd8ed1ab_1
sqlalchemy=2.0.25=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.32.0.post1=pypi_0
statsmodels=0.14.1=pypi_0
stdlib-list=0.10.0=pypi_0
sympy=1.12=pypi_0
syncer=2.0.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
texttable=1.7.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.2=pypi_0
tk=8.6.13=h5083fa2_1
tokenizers=0.15.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.3.3=py310h2aa6e3c_1
tox=4.12.1=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
transformers=4.37.1=pypi_0
types-requests=2.31.0.20240125=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
umap-learn=0.5.5=pypi_0
unidecode=1.3.8=pypi_0
uptrace=1.22.0=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.25.0=pypi_0
virtualenv=20.25.0=pypi_0
watchfiles=0.20.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websockets=12.0=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.9=pypi_0
wrapt=1.16.0=pypi_0
wsproto=1.2.0=pypi_0
xmltodict=0.13.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h965bd2d_0
zipp=3.17.0=pyhd8ed1ab_0
